re-arranged backend files + set up front-end files
feel free to move the scripts folder back but i think it should stay separate for front-end and backend and not be in the parent directory. it was still broken so i fixed it.

set up a venv so any new packages can be frozen into requirements easily. will start a new branch for front-end stuff after this